// models/plant.ts
export interface PlantData {
  category: string
  id: number
  name: string
  scientificName: string
  description: string
  soilId: number
  spacingId: number
  requirementsId: number
  feedingId: number
  stakingId: number
  daysToHarvestMin: number
  daysToHarvestMax: number
  yieldPerPlantMin: number
  yieldPerPlantMax: number
  consumptionChildKg: number
  consumptionAdultKg: number
  storage: string
  image: string
  image2: string
  image3: string
  image4: string
  soilType: string
  soilPreparation: string
  spacingRowCm: string
  spacingPlantCm: string
  requirementsSun: string
  requirementsWater: string
  requirementsGermination: string
  feedingSchedule: string
  stakingRequired: boolean
  stakingNotes: string
  preservation: string
  recipeIdeas: string
}